requirements.txt.txt
streamlit
mysql-connector-python
pandas
plotly
plotly-express